Winning model
Team Name: I.Guyon, L.Sun, D.Kalainathan
Method Name:  Linear Kernels

Method Description
------------------
No train.
For each pixel in a image, compute its correlation with the neighboring pixels at time t-1.
After obtaining a correlation matrix for each pixel, the matrix is rescaled in for the prediction
between frames t-1 and t to be accurate.
Then, the prediction is made with the matrixes and the frame t.
The prediction is made once and copied over all the number of frames to predict.

Hyper-parameters
----------------
- Number of frames used to compute the matrixes weights
- Size of the weight matrixes